Added first three core features
