--- tic-tac-toe/game.js
@@ -1,61 +1,0 @@
-import React, { useState } from 'react';
-import Board from './Board';
-
-<Board
-  squares={current.squares}
-  onClick={(i) => this.handleClick(i)}
-/>
-
-function Game() {
-  const [history, setHistory] = useState([{ squares: Array(9).fill(null) }]);
-  const [stepNumber, setStepNumber] = useState(0);
-  const [xIsNext, setXIsNext] = useState(true);
-
-  const handleClick = (i) => {
-    const newHistory = history.slice(0, stepNumber + 1);
-    const current = newHistory[newHistory.length - 1];
-    const squares = [...current.squares];
-
-    if (calculateWinner(squares) || squares[i]) {
-      return;
-    }
-
-    squares[i] = xIsNext ? 'X' : 'O';
-
-    setHistory([...newHistory, { squares }]);
-    setStepNumber(newHistory.length);
-    setXIsNext(!xIsNext);
-  };
-
-  const current = history[stepNumber];
-  const winner = calculateWinner(current.squares);
-  const status = winner
-    ? `Winner: ${winner}`
-    : `Next player: ${xIsNext ? 'X' : 'O'}`;
-
-  const moves = history.map((step, move) => {
-    const desc = move ? `Go to move #${move}` : 'Go to game start';
-    return (
-      <li key={move}>
-        <button onClick={() => setStepNumber(move)}>{desc}</button>
-      </li>
-    );
-  });
-
-  return (
-    <div className="game">
-      <div className="game-board">
-        <Board
-          squares={current.squares}
-          onClick={(i) => handleClick(i)}
-        />
-      </div>
-      <div className="game-info">
-        <div>{status}</div>
-        <ol>{moves}</ol>
-      </div>
-    </div>
-  );
-}
-
-export default Game;
